+++ socialForcesAI/src/SocialForcesAIModule.cpp
@@ -53,7 +53,7 @@ namespace SocialForcesGlobals
 	float sf_wall_a;
 	float sf_max_speed;
 	float sf_preferred_speed;
-	float sf_furthest_local_target_distance;
+	int sf_furthest_local_target_distance;
 
 	PhaseProfilers * gPhaseProfilers;
 }
